internal/server/singleprocess: send opened confirmation message on exec
This sends a message to the entrypoint when an exec session is successfully
started. This is used to deflake our tests and removes a test that would
block on my machine like 50% of the time. I ran the test 1000 times on
my machine without a failure after this patch.

The CEB itself ignores this message and it is unused currently. We could
use this message in the future as a timeout mechanism for starting an
exec session. If the opened message isn't received in sufficient time,
we could end our exec session.