internal/core: add proper subtype to Any values
Without this, subsequent operations such as push and deploy would get an
error of "type cannot be satisfied". We need to tag our *any.Any values
with the proper subtype otherwise argmapper doesn't know WHAT any.Any
is.

This is paired with a small bug in argmapper that was fixed as well
where named subtype values couldn't satisfy typed subtype values.

The tests for this are covered in argmapper.