sdk: deployment and releaser funcspec should not be a mapper
If `spec.Result` is non-nil, then the funcspec package treats the
generated function as a "mapper" so it returns a fully typed struct so
that the function can be used in a chain of converters.

By specifying `spec.Result` to nil, the original results of the callback
function are preserved, because we expect this to NOT be a mapper (or
rather: to be a target function) and we want the real original results.

This is non-sensible behavior for a single function, so I'm going to
follow up with a change to change this behavior into two separate
functions.