Increase K8s install timeout
AKS takes a bit longer than the other platforms to spin up due to the way PVCs are handled. As a result, the Waypoint server will consistently not install. In testing, The server is ready at about the 80-90 second mark, however our 1 minute timeout fails the install out. Proposing bumping this up to 2 minutes instead to give necessary time.