Strip terminal titles from FeaturesList
We actually don't want terminal titles here as it's not a text file!  I
called this out to design and it looks like they didn't patch it up.

I kept the title implementation in the component though, because it's
good and certainly something that we'll want to use in the future.