Features list section (Desktop) (#357)
* Features list section

* add aria tags

* Eliminate dark theme terminal

Definitely a good catch that the designs deviated, but I'm pretty
strongly confident this deviation from the design side of things was
accidental and not intentional.  I'll firm up this assumption tomorrow
with the team, but for now, I'll remove.

* Lift up props to homepage

* Strip terminal titles from FeaturesList

We actually don't want terminal titles here as it's not a text file!  I
called this out to design and it looks like they didn't patch it up.

I kept the title implementation in the component though, because it's
good and certainly something that we'll want to use in the future.

* Switch to a stable ID

We don't plan on this components props changing dynamically ever in it's
usage, so switching to a stable ID strategy rather than passing in an ID
to smooth out the interface a bit for the end user.

* Plug in some more content

Co-authored-by: Brandon Romano <BrandonRRomano@gmail.com>