internal/core: destroy deploy with no Deployment is not an error
This can happen in some erroneous scenarios. In this case we just mark
the deployment as destroyed.

Post 0.1 we need to do a much tighter job of managing the physical state
of deploys. Evan and I chatted about this already. But for now this will
band-aid a situation where deploys _cant_ be destroyed even though they
don't exist.