website: add notes about status.hashicorp.com
We want any URL service users to be able to find the status
page, so I put it in the footer and on the URL service page.

Longer term, it'd be great if we could render this dynamically
in the header or similar if that statuspage.io component
has issues.