Add Environment and Secrets to default container
This allows specifying additional environment variables to the default
PORT => 3000 mapping, and specifying secrets to use with the default
container.

I'm still trying to work out what the flow should be for a Waypoint
deployment into ECS, but it seems like we have at the very least a
default container that is created during the build stage.