builtin/aws: use SharedConfig to load ~/.aws/config
Fixes #604

This enables loading ~/.aws/config for assumed role auth.

We should talk to the Terraform about making aws-sdk-go-base less
Terraform-specific so we can share it, since they've solved a lot more
problems that we're going to run into: https://github.com/hashicorp/aws-sdk-go-base

But for now, this will work.