internal/clicontext: fall back to plain files if symlinks don't work
Fixes #625

This is a fallback because this is not an _atomic_ writing solution. On
Windows, it appears this isn't possible to begin with so this is best
effort.

Symlinks generally do work fine on Windows and other systems, but if
they don't for any reason (usually permissions), we fall back to normal
files.