Fix bug with setting task_role_name
The `execution_role_name` is optional, but this is a required parameter,
so is automatically created if it's left blank.

Setting `task_role_name` should also be optional, but since this is not
required it isn't created automatically, and should not be referenced in
the task definition if it's not set.

Fixes issue: https://github.com/hashicorp/waypoint/issues/790