send "EOF" event for stdin through exec session when stdin is closed
This adds "EOF" awareness for stdin to exec so that stdin can be
properly closed on the remote side. Without this, some processes would
hang indefinitely waiting for stdin to close. For example, if you run
`cat | ls` (just normally, not via Waypoint), ls hangs forever since it
is waiting for stdin to close.

For non-interactive stdout and an interactive stdin, we treat stdin as
closed. This seems to align with how other Unix tools treat stdin in
this case (such as ls).

This change requires a full client, server, entrypoint update to work,
but is backwards compatible throughout and just falls back to the same
buggy behavior.

Fixes #826