docs/troubleshooting: Include extra help for troubleshooting deployments
This context was taken directly from the Learn tutorials. Given that
it's useful for investigating deployed applications, and the fact that a
user might miss this context if they didn't follow the Learn guide, it
seems useful to also have this context on the docs where it's more
discoverable for people.