Fix path for server proto when using go generate
Prior to this commit, go generate would run in the dir
`internal/server`, so when it went to look for the server.proto, it
would look inside
`waypoint/internal/server/internal/server/proto/server.proto`. This
commit fixes that by ensuring it goes back to the root dir to find
`server.proto`.