internal/server: fix send-after-close possibility
I could never reproduce this bug nor write a test for it since it was
racey in nature but eyeball-tracing I belive this is the fix.

Prior to this change, there was a scenario that could exist where
EntrypointExecStream would exit before the recv goroutine. When
EntrypointExecStream exited, it would close the event channel that the
recv goroutine sends to. If the recv goroutine simultaneously had a
message buffered, it could try to send to the closed channel.

With this change, we moved the channel close into the recv goroutine.
This way there is never a send-after-close possibility.

The one important detail is that the event channel MUST ALWAYS be closed
when the EntrypointExecStream exits, because closing this channel
signals to the client side (typically a CLI) that they should exit
prematurely. Therefore, I moved the goroutine start up before we send
the Open message. This ensures that in every exit scenario, the
goroutine is running, recv should get an EOF, and we'll close the
channel.