internal/ceb: logger should connect when config is ready, not child
The end user behavioral change here is that the log stream will begin
sending to the server as soon as EntrypointConfig is established. This
means that logs will appear even if the child command is not entering
the "ready to start case", which was happening when poorly behaved
config sources were blocking.

Internally, this  changes our state tracking in the CEB to use bools and a
condition var. Previously we were using an atomic+channel. This changes the
requirement to one shared cond var + bool per state.

We aren't using a state enum because these are not mutually exclusive
states, we can have both states true at once.