Store API Token in state to prevent re-registering guest account
Prior to this commit, if a user upgraded or brought back up a Waypoint
server, it would re-register itself with Horizon and get a new URL,
leaving the old one behind. This commit fixes that by storing the API
token, and skipping any registering if it exists from prior server
state.