internal/ceb: quiet ceb log disconnection messages
This changes the ERROR message to a DEBUG and adds a new ERROR-level
message only if we failed to reconnect the log stream.

It is expected that long-running streams such as the log stream will be
disconnected periodically due to any number of reasons, but most
commonly a load balancer in the middle limiting connection times. In
this scenario, we quickly reconnect and its not an issue. It creates too
much noise to log every disconnect as an error.

However, if we fail to reconnect the log stream, it is an error that
should be noted.