internal/runner: step IDs must increment across multiple stepgroups
Previously, the step ID was calculated based on `len(sg.steps)` which
ensured that it was always unique for each new step. But this is only
true as long as there is only one step group ever created.

In cases where multiple step groups are created, this caused old steps
to be reused and it caused some confusing terminal behavior (old lines
would be overwritten as the UI framework did its best to update them).

This moves the step ID counter to the UI level so it is guaranteed to be
monotonically increasing for all steps for that UI.