internal/server/singleprocess: Handle race between exit and cancel
This laces in additional logic to handle a very specific situation: the
client side of the entrypoint stream sends an exit message and then
cancels it's context. In that case, it's possible to recieve the exit
message and have ctx.Done() be closed at the basically the same time.

So we need to not drop the final message on the floor in this case, so
we perform additional work to try and deliver it.