internal/client: use given context so it can be cancelled
While attempting to connect to a server, Ctrl-C was not working. This is
because we were using `context.Background` rather than the passed in
context when attempting to connect.

I don't know why we did this, it looks like a simple mistake from dev.
Verified that this makes Ctrl-C work properly for server connection.