serverconfig: store platform in server config
Prior to this commit, waypoint had no way internally to figure out what
kind of platform a server was, which meant it had to rely on the user to
pass the proper flag when doing any server operations on an existing
context. It also meant it was difficult to realize what a server
platform was when looking at existing context through `waypoint context
list`. This commit updates that to store the context passed in from
install.