internal/server/execclient: Fix tty bug
In the case of `date | waypoint exec` the stdin is a pipe but the stdout
is a tty. We check stdout as a tty because that's what things normally
care about, as they want to know how to display things.

The bug cause us to try to put stdin into raw mode, which it can not
because it's a pipe in this case.