internal/cli: Add ability to start an exec plugin for a session
The semantics here are a bit simplicist but they should work. Basically
if there are no instances for a deployment, we try to generate one by
using an exec plugin. For platforms that don't have an exec plugin, it
will simply error out.

A better way to implement this would be to check if the platform
supports an exec plugin, and if so only use it. But checking for this
fact requires booting the plugin, and booting the plugin requires the
runner config, and thus it's just easier to do it this more simplistic
way.