internal/install: Attempt to use default context for auth token
If a server has been restarted with the install step, the bootstrap
token step will certainly fail due to the server already being
configured. This commit fixes that by attempting to use the default
context to see if its auth token is valid for the server. If not, fail
like it was going to fail. If it's valid, go ahead and use the default
context since its the one we're looking for for the restarted server
container.