internal/{ceb,core,server/singleprocess}: Fix exec stream closure
This fixes a very specific edgecase where if the virtual side of the
exec handler sends an Exit message and then cancels the context in quick
succession, the server side may not see the message, only the context
cancelation. gRPC seems to prioritize the cancelation above pending
events, possibly because message sends are not guaranteed (the docs
state this).

The change then is to wait for the server side to close the stream after
the virtual side sends it's exit code. This means that the server side
is in control of seeing the messages and shutting down the stream.

We also can now simplify the handling of the ExecStream messages by
allowing server.Recv() to return any buffered messages even if the
context has already been canceled. Effectively, if we check that the
context the server stream is using has been canceled on our own, we
don't allow the server stream to return any buffered messages.

The entrypoint exec code is also simplified by having it depend only on
the execclient/exec side by checking exec.Context for readiness when
trying to relay messages back to the execclient.