internal/server/state: project put should not modify applications
We document that project-related APIs cannot be used for applications,
but previously we didn't preserve existing applications. If a user
called UpsertProject to modify an existing project it'd blow away all
their applications.

This couldn't be hit in the real world unless the user was using the API
directly, since `waypoint init -update` is the only thing that calls
this API and it always registers applications again.

But with future changes such as introducing server side waypoint.hcl and
future UI changes to modify projects, this would've become a big issue.