internal/server: QueueJob does not validate the application
We removed this check because it introduces a chicken/egg problem with
onboarding a new project: the user needs to have the waypoint.hcl loaded
to queue a job (to know what apps are defined), but with remote GitOps
style builds, we need to queue a job first to load the waypoint.hcl.

By removing this check, we move to a "late binding" approach where
non-existent applications will still error, but later in the process
after the job starts running.