internal/ceb: Add SSH server for exec plugins
The CEB can be, through the setting of a few environment variables,
configured to listen for a single SSH connection. It will then exec a
single command send by the SSH client, and then exit.

This can be used by platform exec plugins that wish to create one-off
infrastructure to SSH into to provide the shell for a deployment.