builtin/aws/ecr: Return an ECR image rather than Docker
By returning an ECR specific Image, we can allow other plugins that only
support reading images from ECR (like Lambda) enforcing this quality by
only accepting the ECR Image type rather than the Docker image type.

This is one the hallmarks of the argmapper system and this particular
case is one of the reasons it was added to waypoint.