builtin/k8s: Introduce Ports map for multi-port applications
This commit introduces a new config option for k8s: ports.

It allows users to define multiple listening ports that their
application miht need for its container. It also allows users to give
their ports a name, as well as a few other options that a contianer port
might need to be configured with.