builtin/k8s: Surface pod failures on deploy
Prior to this commit, if a pod failed to pull an image, the error
wouldn't surface to the user and it would just look like the deployment
was hanging. This commit fixes that by using the `ui` to surface the
known pod failure of not being able to pull a container image.