internal/datasource: Git polling respects Ref configuration
The previous polling implementation only looked at HEAD. Users may
configure a separate Ref for their Git datasource, but the polling
didn't respect this.

This commit changes the polling to respect the Ref setting. The ref
setting can be in any format `git rev-parse` takes but only hash
references; caret refs and so on such as `HEAD~2` are not allowed.