Introduce go-changelog to Waypoint
This commit adds some essential template files for working with
`go-changelog`. It also includes a basic readme for how to generate the
proper CHANGELOG entries so that they can be generated with the tool.
Finally, it introduces a github workflow for checking when pull requests
forget to include a changelog file, with the option of being ignored if
the proper label is applied to the PR.