internal/core: render all output variables for destroy phases
This builds on #1254 which wasn't quite complete it turns out...

We need to also ensure that we render all the variables for workspace
destruction and "destroy all" operations.

This also changes the behavior slightly so that we render the Waypoint
config using variables based on the _latest_ deploy or release.
Previously, we'd use the deploy/release we were destroying. The issue is
that the configuration we have at the moment is only the current config,
so you could get waypoint.hcl parse errors if they used variables that
didn't exist at one point or another.

Practically, I'd like to update the SDK docs to note that during a
destroy operation you only have the config for the current config rather
than a point-in-time configuration. You should persist all state you
need onto your deployment/release structure. I think that's how all
plugins work anyways today but I'll update docs there to be explicit.