k8s: add more options, rename PORT env to WAYPOINT_EXPOSED_PORT
This commit adds the possibility to specify new config
elements in the HCL file. With these additions is now possible
to define a `container` block that can be used to specify:

- `command`
- `args`

In addition to this change, PodSecurityContext now supports
the `fs_group` option.

This commit additionally adds a (maybe controversial?)
change. The `PORT` environment variable is renamed to
`WAYPOINT_EXPOSED_PORT` to avoid clashing with other environment
variables that might have been specified in `static_environment`