serverinstall/nomad: Provide config values for memory and CPU
This commit introduces config options to use when defining a Server and
Runner nomad job config. If no option is specified, it will use the
configured defaults for Waypoint. This commit also increases the default
CPU in MHz to 200 and Memory in MB to 600 (double the default for Nomad
tasks)

Fixes #1302