cli/{deploy,release}: Account for inplace deploys for release
This commit fixes a bug where the release subcommand would print a
message about in place releases and generations. This is only true if
the deployment is mutable. It also looks to see if the geration id is
empty rather than if the whole struct is nil, since the value ends up
initialized but empty even with immutable deploys