cli/upgrade: Check for required upgrade flags and print better message
Because `-auto-approve` and `-platform` are currently required flags for
using the server upgrade command, this commit modifies how we let the
user know about missing flags to prevent them from including one,
running the command, and seeing another is required. Doing both up front
should reduce the back and forth for when someone wants to run the
upgrade command. If neither are included, we also print the help text
for the upgrade command.

Fixes #1371