plugin/docker: Report on deploy status
This commit introduces the implementation of the Status function for
docker on deploy. It queries the current deployed containers health and
looks for if it was configured using dockers built in health checks. If
not, the function attempts to make some basic assumptions for
determining the health of the container.