Install the correct version of mockery.
Without specifying the package, nothing gets installed.

Without injecting the version via ldflags, mockery prepends the wrong version number (v1.0.0) to all of the mocks that it generates. I do not love hard-coding 1.1.2 in here, but this seems unlikely to change frequently and I don't see another path. I was hoping to use some a bash subcommand to extract the version with something like the following, but it doesn't work with go install and is kind of hacky also.

```
//go:generate go install "-ldflags=-s -w -X github.com/vektra/mockery/mockery.SemVer=$(go mod graph | grep 'github.com/hashicorp/waypoint github.com/vektra/mockery' | awk -F '\@v' '{print $2}')" github.com/vektra/mockery/cmd/mockery
```