CLI list deployments shows health for each entry
In writing this, I realize that our ListStatusReports endpoing lists _all_ status reports for all resources, with no way to filter. We'll probably need to enhance that pretty soon as we build up more report types and volume.

Also it looks like this is a big protobuf diff. The existing generated code has "unknown" protobuf version, so i'm not sure what version I need to revert to to make it match.