Unimplemented status report "Unknown Status" -> "n/a"
Previously, if a plugin does not return status reports, we would say "Unknown Status", which is a bit wordy and also conflicts with the "UNKNOWN" official status, where a plugin _has tried_ to get status and is explicitly returning unknown.

"n/a" kind of implies that the very concept of health doesn't apply to some deployments, which isn't right either, but it's the least-worst thing I can think of.