Moving vendor to 3rdparty
If there is a `vendor` folder at the project root, the Goland IDE ignores go modules and downloads all dependencies into that vendor folder. Renaming `vendor` to `3rdparty` also makes us more consistent with waypoint-plugin-sdk.

I also added tests to the `make gen` commands to verify that the submodule has been initialized, and instructs you to do so if not.