Dockerfile: Remove SSH key injection from circle-ci and dockerfile build
This commit removes the injected ssh keys from the build process as well
as the various circle-ci jobs. It was originally added due to Waypoint
being a private repo and needing access to push the built container to
the alpha releases on github. Waypoint is no longer private, and now circle-ci
should still be able to push packages to github using the docker login step
instead.