config/funcs: template functions to return valid ArgError values
The template functions are all defined with a variadic parameter to
collect zero or more "vars" maps, but that means we need to be careful to
handle the case where there are no vars maps at all.

This code seems to have originally been derived from Terraform's
"templatefile" which has a non-variadic "vars", and so Terraform was
correct to assume that argument 1 would always be present. Since
Waypoint's version allows omitting it, we'll handle that situation by
putting the "blame" on the template argument instead, to suggest that it's
invalid to refer to any variables inside a template call that doesn't have
any variables.

The previous incorrect result was originally a panic due to violating the
contract HCL expects for function.ArgError, although later versions of
HCL handle it by just returning a less-specific error message in that
case. This fix therefore boils down to just highlighting a more specific
part of the input in the error message HCL will ultimately return.