builtin/docker: pass build args into img in correct format
We were previously constructing a string, but `exec/cmd` takes a LIST of
arguments that it passes directly to the exec syscall. The easiest way
to do this is to pass the build args config to buildWithimg and let that
func do its own data formatting to call it correctly.