Enable status report polling after a status report is first generated
This commit updates the status report operation to enable status report
polling on the application that the report was just generated on. If
polling is already enabled, it does nothing. Otherwise it upserts into
the application that polling should be enabled. Doing so means the
queued application poll handler picks up on this change and starts
polling for status reports on the default interval