Not embedding hclConfig as a pointer.
This addresses the panic within https://github.com/hashicorp/waypoint/issues/1704.

The problem here is that, if you embed a pointer to a private struct, and the private struct is nil, the caller has no way to determine if the embedded struct's values are safe to access.

This manifests here: https://github.com/hashicorp/waypoint/blob/main/internal/cli/base.go#L253

There is no way to determine of c.cfg.Runner is safe to access.

A simplified example of the problem:

https://play.golang.org/p/sbAPT2HGJqW

I believe so - we don't have any methods on hclConfig that require it to be a pointer. We also never statefully maniuplate it outside of the context of `cfg`, as far as I can see. I've tested this locally and it doesn't seem to break anything.