internal/server: harden git path validation
This adds more validation to ensure that the path specified for a git
repository is what we expect. There are no bugs I've seen related to
this and users seem to be using subpaths fine, but working on #1821 made
me realize the need to be confident our paths are normalized.