internal/server: disallow deleting the default user
Runners and deployments currently use the default user for API calls
(Runner or Entrypoint API). We need to support the concept of a "service
account" for each of these that we don't currently support today.

For now, we'll prevent deleting this user. You can update the user if
you want cause all that matters is the ID and the ID is immutable.