Introduce GeneratePollJobs
This commit introduces a new poll handler function GeneratePollJobs,
which is expected to return an array of QueueJobRequests. Instead of
queueing a single job, Waypoints queued poller now can queue a slice of
jobs in the case where a poll item has multiple things to generate poll
job requests for.