ui: de-lint `routes/.../project/apps.ts`
Note we remove the `afterModel` hook. If a project contained only one
app then this hook would redirect directly to said app, skipping the
list.  However, it looks like the code had fallen out of step with the
model hook (as had the type annotations) so the predicate would never
match and the redirect would never be performed.

Seeing as the behavior has been inoperative for a while, I chose to
remove it entirely rather than reinstate it.