feat: add status and status-check, improve StatusReportList
This commit adds a `waypoint status` and `waypoint status check`
command intended to provide more insights on the project.

Currently, this commit only shows the status of the deployments
(health check), in a similar way as `waypoint deployment list`
and adds a `status check` command to trigger an Health Check
on the deployments.

In the future, this `status` command should display the
project associated resources and include the releases.

The `status check` command should in the future also include
releases, and let the user choose which check to perform.
In addition to this, the check should be done remotely on the
runner.

This commit also fixes an issue with the list of status reports
being returned by the StatusReportList method: when multiple
checks were performed, the UI wasn't always picking up the
correct StatusReport (most recent). We're now handling
this case in the StatusReport method itself, so that we
only return one StatusReport per Deployment.

Related Hashicorp Waypoint issues:
- hashicorp/waypoint#1501