internal/server: Respect TOS flag for using URL service
Prior to this commit, the TOS for using the URL service in Waypoint
would always be true, regardless of the request. This commit fixes that
by respecting the argument passed through.

Note that the CLI server_run would error out before this reuqest if the flag
was false, so we were not auto-accepting despite the function always setting
the terms to true.

Fixes #1843