internal/server: Ensure ExpediteStatusReport sets a default workspace
Prior to this commit, if a caller requests a status report job but
doesn't specify a workspace, the server would panic on the Workspace ref
being unset when it attempted to build the job. This commit fixes it by
defaulting to the default workspace if the caller does not specify which
workspace the report should run in.