cli/uninstall: Add more validation to requested platform for uninstall
Prior to this commit, the CLI uninstall command assumed that the
platform requested matched the server platform. This could be confusing
if a user has both a docker platform server and a local server from
`waypoint server run`. This commit adds some extra validation with the
platform requested by loading the server platform and taking into
account what platform was requested.

This commit also changes the hard requirement that the uninstall
command needs a platform flag. If a server context has a platform set,
and no platform flag was requested, use that platform instead. Note
that the platform flag, if requested, always overrides what the server
context has. This is because some early versions of Waypoint do
not set the Platform context. If the flag and server platform conflift,
we'll print a note to say we're honoring the platform flag requested.

Fixes #2046