cli/status: Introduce a `-refresh` flag to recalculate status reports
This commit introduces a new flag, `-refresh`, which will refresh all of
the requested status reports for the current view from running `waypoint
status`. This means if you refresh a refresh for a project view, all of
that projects applications will be refreshed. If you request a single
app, only that app will be refreshed, etc.