ui: fix potential bug in `ApiService['upsertProject']`
I wish you could create protobuf messages from objects like this:

    new Application({ name: 'test' })

But alas, that doesn’t work. We can verify by running this snippet in
the browser console:

    (() => {
      const { Application } = require('waypoint-pb').default;
      let myApp = new Application({ name: 'test' });
      console.log(myApp.toObject());
    })()

Which prints this:

    {project: undefined, name: "", fileChangeSignal: ""}

Sadly, `name` is empty.

This leaves us needing to write our own `*FromObject` functions, which
is a little tedious but I guess part of the joy of using protobufs in
JS.