Upgrading plugin SDK version to latest for the 0.5.x branch.
There are bugfixes that we need in here. The only wrinkle is that the ODR additions
were breaking changes to the API, so I also needed to modify our ClientConfig in core
waypoint to disable ODR. I think this is fine, as we're not planning to backport
any of the ODR changes.