ui: fix `flashMessages.error` regression
So it turns out that if you declare an additional field like so:

```ts
class MyClass extends BaseClass {
  myField!: SomeType;
}
```

Then the compiler adds a line to the initializer that defines the property:

```js
class MyClass extends BaseClass {
  constructor(...args) {
    super(...args);
    Object.defineProperty(this, "myField", { value: void 0 });
  }
}
```

Notice that the `Object.defineProperty` comes _after_ the call to `super()`. In
the case of ember-cli-flash, it’s `super()` that the does the work of
registering custom convenience methods, so `super` happily defines `error`, and
then the line added by typescript overwrites it with `void 0` and we end up with
no `error` method to call.