ui: fix types for router service
It turns out my mistake on the first attempt at this was that you
*cannot* merge class declarations. You can, however, merge interfaces
with classes, so that’s what this solution does. This is actually all
documented beautifully on [the declaration merging page].

Note I also moved these declarations out into their own file. This is
because we need the top-level imports, which cause problems if they’re
present in `global.d.ts`.

[the declaration merging page]: https://www.typescriptlang.org/docs/handbook/declaration-merging.html