Dont fail if API service account doesn't have access to list all pods
Prior to this commit, when we went to check for a metrics-server, if the
user account that Waypoint had didn't have access to list pods for every
namespace, the request would fail and return an error. Since this is
mostly for a helpful warning message, we'll log the error instead and
move on.