internal/clicontext: default to gRPC port for login addr
Before we would not set a port at all, and I'm not sure what port it was
trying to connect to. Given our installation method on all platforms
defaults our gRPC to port 9701, it makes sense to default to port 9701
if none is specified.

This makes the initial installation login experience smoother cause most
installations can just do:

  $ waypoint login <IP or hostname>

And it'll just work, rather than looking up the port, which is most
likely 9701.