Improve documentation for path-based hcl functions
In practice, the file hcl functions require absolute paths, so they are really only useful when combined with an hcl variable that gives a relative path (like path.app). I struggled for a while to find a way to make `filebase64` useful before stumbling into the answer.

This will help point users in the right direction.