Increase the height of the left intro terminal
Doing this to roughly match the size of the right terminal for better
layout consistency.

I first tried to shrink the right terminal and add scrolling, but I then
realized that each terminal content size is determined by the largest
tab, so when I added overflow: scroll there it actually was causing the
tabs with little content to unnecessarily scoll.