Split telemetry setup into create and run phases.
This allows us to better surface initialization errors. Thanks for the feedback Evan and Clint!

Note: I took a few stabs at the implementation here. I wanted to try to avoid creating static wrapper structs for each type of opencensus exporter, so I tried to declare them anonymously. This implementation feels ok to me, but it very well might be too clever/esoteric, and i'm open to just doing it the longform way if anyone hates this.