builtin/k8s: Set hostname on release URL if requested in Waypoint
Prior to this commit, the Ingress resource create func would only set
the release URL based on the configured load balancer from k8s. This
commit changes that by first looking to see if the user requested a Host
to be set and use that instead, otherwise try to get the URL from the
configured load balancer.

Fixes #2475